LRTC_SCAD_FanWeiKai/requirements.txt
numpy
scipy
ipykernel
matplotlib
python>=3.9
pytorch>=2.0